chore(architecture): materialize signed AR-2 residual closeout (#727)
Signing-only materialization of fully qualified AR-2 residual source PR #726. Preserve the exact validated tree while creating a GitHub-signed carrier commit for canonical promotion. No content delta or authority expansion.
