+++ README.json
@@ -32,7 +32,7 @@
     "runtime_architecture_refoundation": {
       "status": "AR_2_DOMAIN_EXTRACTIONS_ACTIVE",
       "purpose": "Incrementally refactor the Orchestra Python runtime into bounded clean/hexagonal architecture layers while preserving public behavior and legacy import compatibility.",
-      "human_sources": ["docs/project/ORCHESTRA_RUNTIME_ARCHITECTURE_REFOUNDATION.md", "docs/project/ORCHESTRA_AR2_DOMAIN_CONTEXT_EXTRACTION.md", "docs/project/ORCHESTRA_AR2_DOMAIN_GOVERNANCE_AUTHORITY_EXTRACTION.md", "docs/project/ORCHESTRA_AR2_DOMAIN_CAPABILITIES_CORE_EXTRACTION.md", "docs/project/ORCHESTRA_AR2_DOMAIN_EXECUTION_CORRELATION_EXTRACTION.md", "docs/project/ORCHESTRA_AR2_DOMAIN_EXECUTION_IDENTITY_EXTRACTION.md", "docs/project/ORCHESTRA_AR2_DOMAIN_EXECUTION_LIFECYCLE_CORE_EXTRACTION.md", "docs/architecture/README.md"],
+      "human_sources": ["docs/project/ORCHESTRA_RUNTIME_ARCHITECTURE_REFOUNDATION.md", "docs/project/ORCHESTRA_AR2_DOMAIN_CONTEXT_EXTRACTION.md", "docs/project/ORCHESTRA_AR2_DOMAIN_GOVERNANCE_AUTHORITY_EXTRACTION.md", "docs/project/ORCHESTRA_AR2_DOMAIN_CAPABILITIES_CORE_EXTRACTION.md", "docs/project/ORCHESTRA_AR2_DOMAIN_CAPABILITY_MANIFEST_EXTRACTION.md", "docs/project/ORCHESTRA_AR2_DOMAIN_EXECUTION_CORRELATION_EXTRACTION.md", "docs/project/ORCHESTRA_AR2_DOMAIN_EXECUTION_IDENTITY_EXTRACTION.md", "docs/project/ORCHESTRA_AR2_DOMAIN_EXECUTION_LIFECYCLE_CORE_EXTRACTION.md", "docs/project/ORCHESTRA_AR2_DOMAIN_EXTRACTION_CLOSEOUT.md", "docs/architecture/README.md"],
       "validation_surface": "tests/runtime/test_runtime_architecture_boundaries.py",
       "machine_policy": "machine/governance/runtime-architecture-boundaries.v1.json",
       "machine_policy_schema": "machine/schemas/runtime-architecture-boundaries.v1.schema.json",
@@ -50,8 +50,20 @@
       "governance_authority_validation_surface": "tests/runtime/test_domain_governance_authority.py",
       "capabilities_domain_surface": "orchestra_runtime/domain/capabilities",
       "capabilities_domain_core_surface": "orchestra_runtime/domain/capabilities/core.py",
+      "capabilities_manifest_domain_surface": "orchestra_runtime/domain/capabilities/manifest.py",
       "capabilities_legacy_compatibility_surface": "orchestra_runtime/capabilities.py",
       "capabilities_domain_validation_surface": "tests/runtime/test_domain_capabilities_core.py",
+      "capabilities_manifest_validation_surface": "tests/runtime/test_domain_capability_manifest.py",
+      "governance_kernel_domain_surface": "orchestra_runtime/domain/governance/kernel.py",
+      "governance_kernel_legacy_compatibility_surface": "orchestra_runtime/governance_kernel.py",
+      "governance_kernel_validation_surface": "tests/runtime/test_domain_governance_kernel_contracts.py",
+      "preexecution_policy_domain_surface": "orchestra_runtime/domain/governance/preexecution.py",
+      "preexecution_legacy_compatibility_surface": "orchestra_runtime/preexecution.py",
+      "preexecution_policy_validation_surface": "tests/runtime/test_domain_preexecution_policy.py",
+      "workflow_domain_surface": "orchestra_runtime/domain/orchestration/workflow.py",
+      "workflow_legacy_compatibility_surface": "orchestra_runtime/workflow_contracts.py",
+      "workflow_validation_surface": "tests/runtime/test_domain_workflow_contracts.py",
+      "ar2_residual_closeout_document": "docs/project/ORCHESTRA_AR2_DOMAIN_EXTRACTION_CLOSEOUT.md",
       "execution_domain_surface": "orchestra_runtime/domain/execution",
       "execution_correlation_domain_surface": "orchestra_runtime/domain/execution/correlation.py",
       "execution_correlation_legacy_compatibility_surface": "orchestra_runtime/correlation.py",
