+++ orchestra_runtime/domain/capabilities/__init__.py
@@ -7,12 +7,14 @@ from .core import (
     evaluate_capability_grants,
     intersect_capability_grants,
 )
+from .manifest import RuntimeCapabilityManifest
 
 __all__ = [
     "CapabilityDecision",
     "CapabilityReasonCode",
     "RuntimeCapability",
     "RuntimeCapabilityGrant",
+    "RuntimeCapabilityManifest",
     "enforce_capability_decision",
     "evaluate_capability_grants",
     "intersect_capability_grants",
