+++ orchestra_runtime/domain/governance/__init__.py
@@ -11,15 +11,43 @@ from .authority import (
     TargetSelector,
     TargetSelectorType,
 )
+from .kernel import (
+    GOVERNANCE_KERNEL_SCHEMA_VERSION,
+    ArbiterKernelResult,
+    ArbiterReasonCode,
+    GovernanceDecision,
+    GovernanceDecisionRecord,
+    TransitionDisposition,
+)
+from .preexecution import (
+    PREEXECUTION_SCHEMA_VERSION,
+    ExecutionAction,
+    ExecutionIntent,
+    PreExecutionConstraint,
+    PreExecutionPolicy,
+    PreExecutionReason,
+)
 
 __all__ = [
+    "GOVERNANCE_KERNEL_SCHEMA_VERSION",
+    "PREEXECUTION_SCHEMA_VERSION",
+    "ArbiterKernelResult",
+    "ArbiterReasonCode",
     "AuthorityDecision",
     "AuthorityProvenance",
     "AuthorityReasonCode",
     "AuthorityScope",
     "Constraint",
     "ConstraintKind",
+    "ExecutionAction",
+    "ExecutionIntent",
+    "GovernanceDecision",
+    "GovernanceDecisionRecord",
+    "PreExecutionConstraint",
+    "PreExecutionPolicy",
+    "PreExecutionReason",
     "ProvenanceSource",
     "TargetSelector",
     "TargetSelectorType",
+    "TransitionDisposition",
 ]
