+++ orchestra_runtime/domain/orchestration/__init__.py
@@ -1,0 +1,5 @@
+"""Pure orchestration-domain contracts."""
+
+from .workflow import WORKFLOW_SANITY_SCHEMA_VERSION, WorkflowSanityReceipt
+
+__all__ = ["WORKFLOW_SANITY_SCHEMA_VERSION", "WorkflowSanityReceipt"]
