+++ tests/runtime/test_domain_governance_kernel_contracts.py
@@ -1,0 +1,114 @@
+from __future__ import annotations
+
+import ast
+import inspect
+
+import pytest
+
+from orchestra_runtime import governance_kernel as legacy_kernel
+from orchestra_runtime.domain.governance import kernel as domain_kernel
+
+
+def test_legacy_governance_contract_symbols_are_domain_symbols() -> None:
+    assert legacy_kernel.GovernanceDecision is domain_kernel.GovernanceDecision
+    assert legacy_kernel.TransitionDisposition is domain_kernel.TransitionDisposition
+    assert legacy_kernel.ArbiterReasonCode is domain_kernel.ArbiterReasonCode
+    assert legacy_kernel.GovernanceDecisionRecord is domain_kernel.GovernanceDecisionRecord
+    assert legacy_kernel.ArbiterKernelResult is domain_kernel.ArbiterKernelResult
+
+
+def test_domain_governance_contracts_preserve_normalization_and_digest() -> None:
+    record = domain_kernel.GovernanceDecisionRecord(
+        reviewer=" arbiter ",
+        project_context=" unit-1 ",
+        decision="APPROVED",
+        reason=" validated ",
+        risks=(" none ",),
+        evidence_refs=(" evidence:1 ",),
+    )
+    assert record.reviewer == "arbiter"
+    assert record.project_context == "unit-1"
+    assert record.decision is domain_kernel.GovernanceDecision.APPROVED
+    assert record.reason == "validated"
+
+    result = domain_kernel.ArbiterKernelResult(
+        disposition=domain_kernel.TransitionDisposition.AUTO_CONTINUE,
+        reason_codes=(domain_kernel.ArbiterReasonCode.CONTINUATION_READY,),
+        input_digest="abc",
+    )
+    assert result.to_dict()["disposition"] == "AUTO_CONTINUE"
+    assert len(result.digest) == 64
+    result.assert_claimed_disposition("AUTO_CONTINUE")
+
+
+def test_domain_governance_contract_fail_closed_edges() -> None:
+    with pytest.raises(TypeError, match="must be a string"):
+        domain_kernel._nonempty(1, "field")  # type: ignore[arg-type]
+    with pytest.raises(ValueError, match="non-empty"):
+        domain_kernel._nonempty(" ", "field")
+    with pytest.raises(ValueError, match="unsupported decision"):
+        domain_kernel._enum("MAGIC", domain_kernel.GovernanceDecision, "decision")
+    with pytest.raises(ValueError, match="unsupported governance decision schema"):
+        domain_kernel.GovernanceDecisionRecord(
+            reviewer="arbiter",
+            project_context="unit",
+            decision="APPROVED",
+            reason="ok",
+            schema_version="bad",
+        )
+    with pytest.raises(TypeError, match="human_review_required"):
+        domain_kernel.GovernanceDecisionRecord(
+            reviewer="arbiter",
+            project_context="unit",
+            decision="APPROVED",
+            reason="ok",
+            human_review_required=1,  # type: ignore[arg-type]
+        )
+    with pytest.raises(TypeError, match="risks"):
+        domain_kernel.GovernanceDecisionRecord(
+            reviewer="arbiter",
+            project_context="unit",
+            decision="APPROVED",
+            reason="ok",
+            risks=["risk"],  # type: ignore[arg-type]
+        )
+    result = domain_kernel.ArbiterKernelResult(
+        disposition=domain_kernel.TransitionDisposition.AUTO_CONTINUE,
+        reason_codes=(domain_kernel.ArbiterReasonCode.CONTINUATION_READY,),
+        input_digest="abc",
+    )
+    with pytest.raises(ValueError, match="conflicts"):
+        result.assert_claimed_disposition("STOP")
+
+
+def test_domain_governance_kernel_contracts_have_only_inward_runtime_dependencies() -> None:
+    tree = ast.parse(inspect.getsource(domain_kernel))
+    forbidden_stdlib = {
+        "os",
+        "pathlib",
+        "socket",
+        "sqlite3",
+        "subprocess",
+        "shutil",
+        "tempfile",
+    }
+    forbidden_runtime = {
+        "orchestra_runtime.application",
+        "orchestra_runtime.infrastructure",
+        "orchestra_runtime.entrypoints",
+        "orchestra_runtime.machine_contracts",
+        "orchestra_runtime.governance_kernel",
+        "orchestra_runtime.evidence",
+    }
+    for node in ast.walk(tree):
+        if isinstance(node, ast.Import):
+            for alias in node.names:
+                assert alias.name.split(".", 1)[0] not in forbidden_stdlib
+                assert not any(alias.name.startswith(prefix) for prefix in forbidden_runtime)
+        elif isinstance(node, ast.ImportFrom):
+            module = node.module or ""
+            if node.level == 0:
+                assert module.split(".", 1)[0] not in forbidden_stdlib
+                assert not any(module.startswith(prefix) for prefix in forbidden_runtime)
+            else:
+                assert module.startswith("shared.")
